[IMP] l10n_lv: add tax report attachment tags
Add tags for the tax report attachments added in the related
enterprise PR and tag all existing taxes. They determine the transaction type.

It also adds the special tag `Rep` that is irrelevant for the transaction type.
From a base amount line tagged with `Rep` we only put 40% of the amount
for the attachments.

task-4251184

X-original-commit: 5674a2d6cc9d3fd8fc5a715aaa80329fdc23cff1
Part-of: odoo/odoo#272058
Related: odoo/enterprise#121753
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>